test(e2e): compile and lint the Enterprise half of the suite on every push
The Enterprise half of the end-to-end suite was compiled by nothing but a
person running make test-e2e-docker-enterprise by hand. Every *_ee_test.go
carries `e2e && enterprise`, no workflow passed that tag, GO_ANALYSIS_TAGS did
not list it, and so 41 files never met go vet, golangci-lint or a compiler
between commits. A licensed run in CI is ruled out, since the licences
available are 30-day trials; this is the part of issue 570 that needs none.

`enterprise` is not one more entry in the analysis tag list, because it is not
one more suite: it switches test/e2e/suite between two halves that exclude
each other, *_ce_test.go behind `e2e && !enterprise` and *_ee_test.go behind
`e2e && enterprise`. One analysis run sees one half and never the other, so
listing the tag would have traded the 128 CE files for the 41 EE ones.
make golangci-lint and make analyze therefore run a second golangci-lint pass
with the tag added, scoped to that one package, and the CI compile job
compiles the suite a second time with both tags. e2eTags in gen_testing_docs
stays as it is: the generated metrics read each directory and already count
both halves, and the comment now says why the tag is absent there.

The first run of that pass found what nothing had looked at. Nineteen helpers
and four constants only CE tests use sat in files both halves compile, unused
under the Enterprise tag; they move to assert_helpers_ce_test.go and
network_endpoints_ce_test.go, behind the constraint of their callers. In the
EE files themselves: TestGroupDatadogIntegration built its HTTP helpers as
closures without t.Helper() and was over the cognitive-complexity ceiling, so
they become a groupDatadogAPI type; TestMeta_StorageMoves repeated one
refusal check in nine subtests, now requireRoutingRefusal; the LDAP test's
three identical status chains become ldapIntegrationAbsent; two tests shadowed
an outer err inside their subtests; and the licence cache read carries the
gosec annotation the CE suites already use for operator-supplied paths.
