+++ docs/development/testing/testing.md
@@ -18,13 +18,13 @@
 
 | Metric                                                |  Value |
 | ----------------------------------------------------- | -----: |
-| Total test functions                                  | 14,742 |
-| Unit test functions                                   | 14,139 |
+| Total test functions                                  | 14,741 |
+| Unit test functions                                   | 14,138 |
 | E2E test functions                                    |    603 |
-| cmd test functions                                    |  2,821 |
+| cmd test functions                                    |  2,820 |
 | Test files (internal/)                                |    518 |
 | Test files (cmd/)                                     |    191 |
-| Test files (test/e2e/)                                |    238 |
+| Test files (test/e2e/)                                |    240 |
 | Tool sub-packages tested                              |    177 |
 | Core packages tested                                  |     22 |
 | Overall coverage (`go test ./internal/... ./cmd/...`) |  98.3% |
@@ -35,7 +35,7 @@
 
 | Pattern                                |  Count |     % |
 | -------------------------------------- | -----: | ----: |
-| `TestFunc_Scenario` (2-part)           | 11,837 | 80.3% |
+| `TestFunc_Scenario` (2-part)           | 11,836 | 80.3% |
 | `TestFunc` (no underscore)             |    981 |  6.7% |
 | `TestFunc_Scenario_Expected` (3+ part) |  1,924 | 13.1% |
 
@@ -48,9 +48,9 @@
 | Core packages           |          2,426 |        144 | shared runtime packages such as config, GitLab client, OAuth, resources, prompts, and utilities |
 | Tools orchestration     |            332 |         15 | registration, meta-tool dispatch, safe mode, validation, markdown, and routing tests            |
 | Tool sub-packages (177) |          8,560 |        359 | domain-specific GitLab tool handlers                                                            |
-| E2E integration         |            603 |        238 | build-tagged; only test/e2e/suite and test/e2e/orbit need a real instance                       |
-| cmd packages            |          2,821 |        191 | server entry point and developer command utilities                                              |
-| **Total**               |     **14,742** |    **947** |                                                                                                 |
+| E2E integration         |            603 |        240 | build-tagged; only test/e2e/suite and test/e2e/orbit need a real instance                       |
+| cmd packages            |          2,820 |        191 | server entry point and developer command utilities                                              |
+| **Total**               |     **14,741** |    **949** |                                                                                                 |
 
 ### Core Packages
