+++ test/e2e/README.md
@@ -14,6 +14,8 @@ There are five modules, answering different questions:
 
 Each tag has to be listed in `GO_ANALYSIS_TAGS` in the Makefile and in `e2eTags` in `cmd/gen_testing_docs`, or the module is invisible to `go vet`, to `golangci-lint` and to the generated test metrics. A file behind a tag nothing names is analysed by nothing.
 
+`enterprise` is the one tag that is not a module's. It switches `test/e2e/suite` between two halves that exclude each other: every `*_ce_test.go` carries `e2e && !enterprise` and every `*_ee_test.go` carries `e2e && enterprise`, so a single analysis run sees one half and never the other, and putting the tag in the list above would trade the 128 CE files for the 41 EE ones. It therefore has an analysis pass of its own (`GO_ANALYSIS_ENTERPRISE_TAGS`, scoped to the suite, run by `make golangci-lint`) and a compile step of its own in CI, and it stays out of `e2eTags` because the generated metrics read each directory and already count both halves. Until issue 570 nothing compiled the EE half outside a manual `make test-e2e-docker-enterprise`, which is how nineteen helpers and four constants only CE tests use came to sit unused under that tag in files both halves share; they now live in `assert_helpers_ce_test.go` and `network_endpoints_ce_test.go`, behind the constraint of the tests that use them.
+
 ## HTTP transport module
 
 ```bash
