+++ test/e2e/suite/network_endpoints_test.go
@@ -7,10 +7,8 @@
 package suite
 
 import (
-	"net/url"
 	"os"
 	"strings"
-	"testing"
 )
 
 // Default Docker network endpoints used when setup-gitlab.sh has not written
@@ -39,28 +37,9 @@ func e2eFixtureServiceURL(resourcePath string) string {
 	return joinE2EURL(baseURL, resourcePath)
 }
 
-// remoteMirrorTargetURL returns an authenticated Git URL for target that GitLab
-// can reach from inside the Docker network when configuring push mirrors.
-func remoteMirrorTargetURL(t *testing.T, target ProjectFixture) string {
-	t.Helper()
-
-	token := os.Getenv("GITLAB_TOKEN")
-	requireTruef(t, token != "", "GITLAB_TOKEN is required to configure a remote mirror target")
-
-	baseURL := os.Getenv("E2E_GITLAB_INTERNAL_URL")
-	if baseURL == "" {
-		baseURL = defaultE2EGitLabInternalURL
-	}
-
-	parsedURL, err := url.Parse(baseURL)
-	requireNoError(t, err, "parse internal GitLab URL")
-	parsedURL.User = url.UserPassword("oauth2", token)
-	parsedURL.Path = strings.TrimRight(parsedURL.Path, "/") + "/" + strings.TrimLeft(target.Path, "/") + ".git"
-	parsedURL.RawQuery = ""
-	parsedURL.Fragment = ""
-
-	return parsedURL.String()
-}
+// remoteMirrorTargetURL lives in network_endpoints_ce_test.go: only the CE
+// mirror suites call it, and a helper both halves compile but one half uses
+// is unused under the other's tag.
 
 // joinE2EURL joins baseURL and resourcePath without preserving duplicate slash
 // boundaries from either side.
